+++ dask/array/core.py
@@ -48,6 +48,7 @@ from .numpy_compat import _Recurser, _make_sliced_dtype
 from .slicing import slice_array, replace_ellipsis
 from .blockwise import blockwise
 
+from .tiledb_io import from_tiledb, to_tiledb
 
 config.update_defaults({'array': {
     'chunk-size': '128MiB',
@@ -1782,9 +1783,9 @@ class Array(DaskMethodsMixin):
     def to_tiledb(self, uri, *args, **kwargs):
         """Save array to the TileDB storage manager
 
-        See https://docs.tiledb.io for details about the format and engine.
-
         See function ``to_tiledb()`` for argument documentation.
+
+        See https://docs.tiledb.io for details about the format and engine.
         """
         return to_tiledb(self, uri, *args, **kwargs)
 
@@ -2257,119 +2258,6 @@ def to_zarr(arr, url, component=None, storage_options=None,
     return arr.store(z, lock=False, compute=compute,
                      return_stored=return_stored)
 
-
-def from_tiledb(uri, attribute=None, chunks=None,
-                storage_options=None, **kwargs):
-    """Load array from the TileDB storage database
-
-    See https://docs.tiledb.io for more information.
-
-    Parameters
-    ----------
-    uri: TileDB array or str
-        Location of the data
-    """
-    def tiledb_to_chunks(array):
-        """
-        tiledb tiling:
-            - http://docs.dask.org/en/latest/array-chunks.html
-        dask chunks:
-            - http://docs.dask.org/en/latest/array-chunks.html
-
-        TileDB tile extends corresponds to form 2 of dask chunk
-        specifiers:
-            "A uniform chunk shape like (1000, 2000, 3000),
-             meaning chunks of size 1000 in the first axis,
-             2000 in the second axis, and 3000 in the third"
-
-        :param array: tiledb.Array
-        :return: list
-        """
-        schema = array.schema
-        return list(schema.domain.dim(i).tile for i in range(schema.ndim))
-
-    import tiledb
-    tiledb_config = storage_options or dict()
-    key = tiledb_config.pop('key', None)
-
-    if isinstance(uri, tiledb.Array):
-        tdb = uri
-    else:
-        tdb = tiledb.open(uri, attr=attribute, config=tiledb_config, key=key)
-
-    if tdb.schema.sparse:
-        raise Exception("Sparse TileDB arrays are not supported")
-
-    if not attribute:
-        if tdb.schema.nattr > 1:
-            raise TypeError("keyword 'attribute' must be provided"
-                            "when loading a multi-attribute TileDB array")
-        else:
-            attribute = tdb.schema.attr(0).name
-
-    chunks = tiledb_to_chunks(tdb)
-
-    assert(len(chunks) == tdb.schema.ndim)
-
-    return from_array(tdb, chunks, name='tiledb-%s' % uri)
-
-
-def to_tiledb(darray, uri, compute=True, return_stored=False,
-              storage_options=None, **kwargs):
-    """Save 'array' using the TileDB storage manager, to any TileDB-supported URI,
-    including local disk, S3, or HDFS.
-
-    See https://docs.tiledb.io for more information about TileDB.
-
-    Parameters
-    ----------
-
-    darray: dask.array
-        A dask array to write.
-    uri:
-        Any supported TileDB storage location.
-    storage_options:
-        Dict containing any configuration options for the TileDB backend.
-        see https://docs.tiledb.io/en/stable/tutorials/config.html
-    compute, return_stored: see ``store()``
-    :return:
-        None
-    """
-    import tiledb
-
-    tiledb_config = storage_options or dict()
-    # encryption key, if any
-    key = tiledb_config.pop('key', None)
-
-    if not isinstance(darray, Array):
-        raise TypeError("to_tiledb first argument must be a dask array")
-
-    if not _check_regular_chunks(darray.chunks):
-        raise ValueError('Attempt to save array to TileDB with irregular '
-                         'chunking, please call `arr.rechunk(...)` first.')
-
-    if isinstance(uri, str):
-        chunks = [c[0] for c in darray.chunks]
-        key = kwargs.pop('key', None)
-        # create a suitable, empty, writable TileDB array
-        tdb = tiledb.empty_like(uri, darray, tile=chunks, config=tiledb_config, key=key)
-    elif isinstance(uri, tiledb.Array):
-        tdb = uri
-        # sanity checks
-        if not ((darray.shape == tdb.shape) and
-                (darray.dtype == tdb.dtype) and
-                (darray.ndim == tdb.ndim)):
-            raise ValueError("Target TileDB array layout is not compatible with "
-                             "source array")
-    else:
-        raise ValueError("'uri' must be string pointing to supported TileDB store location "
-                         "or an open, writable TileDB array.")
-
-    if not tdb.isopen and tdb.iswritable:
-        raise ValueError("Target TileDB array is not open and writable.")
-
-    return darray.store(tdb, lock=False, compute=compute,
-                        return_stored=return_stored)
 
 
 def _check_regular_chunks(chunkset):
